third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-cross-document-same-origin.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
async_test(t => {
  let target_url = location.href + "?1";
  appHistory.onnavigate = t.step_func_done(e => {
    assert_true(e.cancelable);
    assert_true(e.canRespond);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    e.respondWith(Promise.resolve().then(
      t.step_func_done(() => assert_equals(location.href, target_url))));
  });

  window.onload = t.step_func(() => location.href = target_url);
}, "AppHistoryNavigateEvent.respondWith intercepts a same-origin cross-document navigation");
</script>
